backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-Mail==0.9.1
APScheduler==3.10.0
python-dotenv==0.19.0
mysql-connector-python==8.0.22
google_auth_oauthlib==0.4.2
google-api-python-client==1.12.8
SQLAlchemy==1.3.22
pytz==2021.3



PS C:cd "c:\Users\cscpr\Downloads\Hospital-Information-System-Cardiology-Department--main\frontend" ; $env:NODE_OPTIONS="--openssl-legacy-provider" ; npm run serve
PS C:cd "c:\Users\cscpr\Downloads\Hospital-Information-System-Cardiology-Department--main\backend" ; python app.py